Fix component and state management
1. Separate App component from index.jsx
2. Use filter method in done button handler
3. Use todoInput state in add button handler
4. Use Date.now() method for Row component unique key
5. Remove unnecessary props in event handler
