+++ src/components/Row.jsx
@@ -1,8 +1,8 @@
 export default function Row({ index, todo, onClick }) {
   return (
-    <p>
+    <div>
       {`${index + 1}. ${todo}`}
       <button type="button" onClick={() => onClick(index)}>완료</button>
-    </p>
+    </div>
   );
 }
